# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Support Ticket Management API",
    "version": "1.0.0",
    "description": "Support Ticket Management — Student Project Assignment\n\nOverview:\nBuild a backend for a company helpdesk system where employees raise tickets, support staff handle them, and managers track everything.\nThe API follows REST principles."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "LoginDTO": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" }
        }
      },
      "CreateUserDTO": {
        "type": "object",
        "required": ["name", "email", "password", "role"],
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" },
          "role": { "type": "string", "enum": ["MANAGER", "SUPPORT", "USER"] }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["MANAGER", "SUPPORT", "USER"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateTicketDTO": {
        "type": "object",
        "required": ["title", "description"],
        "properties": {
          "title": { "type": "string", "minLength": 5 },
          "description": { "type": "string", "minLength": 10 },
          "priority": { "type": "string", "enum": ["LOW", "MEDIUM", "HIGH"] }
        }
      },
      "Ticket": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["OPEN", "IN_PROGRESS", "RESOLVED", "CLOSED"] },
          "priority": { "type": "string", "enum": ["LOW", "MEDIUM", "HIGH"] },
          "createdBy": {
            "type": "object",
            "properties": {
              "_id": { "type": "string" },
              "name": { "type": "string" },
              "role": { "type": "string" }
            }
          },
          "assignedTo": {
            "type": "object",
            "nullable": true,
            "properties": {
              "_id": { "type": "string" },
              "name": { "type": "string" },
              "role": { "type": "string" }
            }
          },
          "comments": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TicketComment" }
          },
          "statusLogs": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/StatusLog" }
          },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "AssignDTO": {
        "type": "object",
        "required": ["userId"],
        "properties": {
          "userId": { "type": "string" }
        }
      },
      "UpdateStatusDTO": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "type": "string", "enum": ["OPEN", "IN_PROGRESS", "RESOLVED", "CLOSED"] }
        }
      },
      "CommentDTO": {
        "type": "object",
        "required": ["comment"],
        "properties": {
          "comment": { "type": "string" }
        }
      },
      "TicketComment": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "authorName": { "type": "string" },
          "authorRole": { "type": "string", "enum": ["MANAGER", "SUPPORT", "USER"] },
          "authorId": { "type": "string" },
          "comment": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "StatusLog": {
        "type": "object",
        "properties": {
          "oldStatus": { "type": "string" },
          "newStatus": { "type": "string" },
          "changedBy": {
            "type": "object",
            "properties": {
              "name": { "type": "string" },
              "role": { "type": "string" }
            }
          },
          "changedAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "tags": [
    { "name": "Auth", "description": "Authentication endpoints (login)" },
    { "name": "Users", "description": "User management (MANAGER only)" },
    { "name": "Tickets", "description": "Create, view, assign, update status, and delete tickets" },
    { "name": "Comments", "description": "Add, list, edit, and delete ticket comments" }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/users": {
      "post": {
        "tags": ["Users"],
        "summary": "Create user (MANAGER)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserDTO" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      },
      "get": {
        "tags": ["Users"],
        "summary": "List users (MANAGER)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/tickets": {
      "post": {
        "tags": ["Tickets"],
        "summary": "Create ticket (USER, MANAGER)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTicketDTO" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Ticket" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      },
      "get": {
        "tags": ["Tickets"],
        "summary": "Get tickets (MANAGER=all, SUPPORT=assigned, USER=own)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Ticket" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/tickets/{id}/assign": {
      "patch": {
        "tags": ["Tickets"],
        "summary": "Assign ticket (MANAGER, SUPPORT)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AssignDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Ticket" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not Found" }
        }
      }
    },
    "/tickets/{id}/status": {
      "patch": {
        "tags": ["Tickets"],
        "summary": "Update ticket status (MANAGER, SUPPORT)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateStatusDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Ticket" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not Found" }
        }
      }
    },
    "/tickets/{id}": {
      "delete": {
        "tags": ["Tickets"],
        "summary": "Delete ticket (MANAGER)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not Found" }
        }
      }
    },
    "/tickets/{id}/comments": {
      "post": {
        "tags": ["Comments"],
        "summary": "Add comment to ticket",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CommentDTO" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TicketComment" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not Found" }
        }
      },
      "get": {
        "tags": ["Comments"],
        "summary": "List comments for a ticket",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/TicketComment" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not Found" }
        }
      }
    },
    "/comments/{id}": {
      "patch": {
        "tags": ["Comments"],
        "summary": "Edit comment (author or MANAGER)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CommentDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TicketComment" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not Found" }
        }
      },
      "delete": {
        "tags": ["Comments"],
        "summary": "Delete comment (author or MANAGER)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not Found" }
        }
      }
    }
  }
}